# swagger/swagger.json
{
	"components": {
		"schemas": {
...
			"malak.Block": {
				"properties": {
					"children": {
						"items": {
							"$ref": "#/components/schemas/malak.Block"
						},
						"type": "array"
					},
					"content": {},
					"id": {
						"type": "string"
					},
					"props": {
						"additionalProperties": true,
						"type": "object"
					},
					"type": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"malak.Update": {
				"properties": {
					"content": {
						"items": {
							"$ref": "#/components/schemas/malak.Block"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string"
					},
					"created_by": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"is_pinned": {
						"description": "If this update is pinned",
						"type": "boolean"
					},
					"metadata": {
						"$ref": "#/components/schemas/malak.UpdateMetadata"
					},
					"reference": {
						"type": "string"
					},
					"sent_at": {
						"type": "string"
					},
					"sent_by": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/malak.UpdateStatus"
					},
					"title": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"workspace_id": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"malak.UpdateMetadata": {
				"type": "object"
			},
...
			"malak.UpdateStatus": {
				"enum": [
					"draft",
					"sent"
				],
				"type": "string",
				"x-enum-varnames": [
					"UpdateStatusDraft",
					"UpdateStatusSent"
				]
			},
...
			"server.APIStatus": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object"
			},
...
			"server.contentUpdateRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"update": {
						"items": {
							"$ref": "#/components/schemas/malak.Block"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"update"
				],
				"type": "object"
			},
...
			"server.createdUpdateResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"update": {
						"$ref": "#/components/schemas/malak.Update"
					}
				},
				"required": [
					"message",
					"update"
				],
				"type": "object"
			},
...
			"server.fetchUpdateReponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"update": {
						"$ref": "#/components/schemas/malak.Update"
					}
				},
				"required": [
					"message",
					"update"
				],
				"type": "object"
			},
...
			"server.sendUpdateRequest": {
				"properties": {
					"emails": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"send_at": {
						"type": "integer"
					}
				},
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"in": "header",
				"name": "Authorization",
				"type": "apiKey"
			}
		}
	},
	"info": {
		"contact": {
			"email": "lanre@ayinke.ventures",
			"name": "Ayinke Ventures"
		},
		"title": "Malak's API documentation",
		"version": "0.1.0"
	},
	"openapi": "3.0.3",
	"paths": {
...
		"/updates/react": {
			"get": {
				"description": "Fetch a specific update",
				"operationId": "reactPost",
				"parameters": [
					{
						"description": "provider type",
						"in": "query",
						"name": "provider",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "email id",
						"in": "query",
						"name": "email_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			}
		},
...
		"/workspaces/updates/{reference}": {
			"delete": {
				"description": "Delete a specific update",
				"operationId": "deleteUpdate",
				"parameters": [
					{
						"description": "update unique reference.. e.g update_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			},
			"get": {
				"description": "Fetch a specific update",
				"operationId": "fetchUpdate",
				"parameters": [
					{
						"description": "update unique reference.. e.g update_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.fetchUpdateReponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			},
			"post": {
				"description": "Send an update to real users",
				"operationId": "sendUpdate",
				"parameters": [
					{
						"description": "update unique reference.. e.g update_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/server.sendUpdateRequest"
							}
						}
					},
					"description": "request body to send an update",
					"required": true,
					"x-originalParamName": "message"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			},
			"put": {
				"description": "Update a specific update",
				"operationId": "updateContent",
				"parameters": [
					{
						"description": "update unique reference.. e.g update_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/server.contentUpdateRequest"
							}
						}
					},
					"description": "update content body",
					"required": true,
					"x-originalParamName": "message"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			}
		},
...
		"/workspaces/updates/{reference}/duplicate": {
			"post": {
				"description": "Duplicate a specific update",
				"operationId": "duplicateUpdate",
				"parameters": [
					{
						"description": "update unique reference.. e.g update_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.createdUpdateResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			}
		},
		"/workspaces/updates/{reference}/pin": {
			"post": {
				"description": "Toggle pinned status a specific update",
				"operationId": "toggleUpdatePin",
				"parameters": [
					{
						"description": "update unique reference.. e.g update_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.createdUpdateResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"updates"
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:5300/v1"
		}
	]
}